requirements.txt
Django
django-material
